+++ packages/camera/camera_android/android/src/main/java/io/flutter/plugins/camera/CameraApiImpl.java
@@ -342,6 +342,11 @@ final class CameraApiImpl implements Messages.CameraApi {
   }
 
   @Override
+  public void setImageQuality(@NonNull Long quality) {
+    camera.setImageQuality(quality);
+  }
+
+  @Override
   public void dispose() {
     if (camera != null) {
       camera.dispose();